Fix pkg description
Per instructions from https://packaging.python.org/guides/making-a-pypi-friendly-readme/,
if we want to have a long_description from a README.md we need to set the
encoding on the read argument (even with python3), and we need to set a long_description_content_type.

With this fix, you could upload it to Pypi and the description would show up as expected.

Add option to ignore pending transaction

When pending transactions become final, they may have their amount, transaction
id, date, etc. re-stated. Some accounting import tools are not smart enough to
avoid duplicate entries that have to be manually pruned later.

This commit adds an option for users who are willing to sacrifice some
up-to-date data in order to avoid this problem entirely, by only looking at
final transactions.